+++ recipes/default.rb
@@ -22,21 +22,22 @@ end
 
 # Download the database if it has changed
 remote_file "/tmp/#{node['squidguard']['blacklist_name']}" do
-  source "#{node['squidguard']['blacklist_url']}"
-  checksum node['squidguard']['blacklist_cksum']
+  source "#{node['squidguard']['blacklist_url']}/#{node['squidguard']['blacklist_name']}"
   user node['squidguard']['user']
   group node['squidguard']['group']
   action :create
-  notifies :run, 'ruby_block[generate_checksum]', :delayed
+  notifies :run, 'ruby_block[generate_checksum]', :immediate
+  notifies :run, 'bash[unpack_and_update]', :immediate
+  only_if days_passed?(node['squidguard']['blacklist_date'], node['squidguard']['update_days'])
 end
 
 # Unpack database, generate checksum, delete temp file
 ruby_block 'generate_checksum' do
   block do
     node.default['squidguard']['blacklist_cksum'] = Digest::SHA256.hexdigest File.read "/tmp/#{node['squidguard']['blacklist_name']}"
+    node.default['squidguard']['blacklist_date'] = date_today
   end
   action :nothing
-  notifies :run, 'bash[unpack_and_update]', :delayed
 end
 
 bash 'unpack_and_update' do
@@ -50,5 +51,4 @@ bash 'unpack_and_update' do
     chown -R #{node['squidguard']['user']}:#{node['squidguard']['group']} #{node['squidguard']['db_dir']}
     EOH
   action :nothing
-  #notifies :restart, 'service[squid]', :delayed
 end